requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.5
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
joblib==1.5.1
keras==3.11.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
networkx==3.5
numpy==2.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
pydicom==3.0.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rich==14.1.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
